requirements.txt
spotipy
python-osc
python-dotenv
psutil
pynvml
